feat: NavBar 추가 및 App.vue 이식
