+++ frontend/package.json
@@ -9,9 +9,11 @@
   },
   "dependencies": {
     "axios": "^1.4.0",
+    "bootstrap": "^5.3.0",
+    "bootstrap-vue-3": "^0.5.1",
     "core-js": "^3.8.3",
     "pinia": "^2.1.4",
-    "vue": "^3.2.13",
+    "vue": "^3.3.4",
     "vue-router": "^4.0.13"
   },
   "devDependencies": {
