+++ frontend/src/components/NavBar.vue
@@ -1,0 +1,93 @@
+<template>
+    <div>
+        <b-navbar toggleable="lg" type="dark" variant="info" fixed="top" sticky="true">
+            <b-navbar-brand href="/">기술 블로그</b-navbar-brand>
+            <b-navbar-toggle target="nav-collapse"></b-navbar-toggle>
+
+            <b-collapse id="nav-collapse" is-nav>
+                <b-navbar-nav>
+                    <b-nav-item href="/board/machine">머신러닝</b-nav-item>
+                    <b-nav-item href="/board/data">데이터분석</b-nav-item>
+                    <b-nav-item href="/board/program">프로그래밍</b-nav-item>
+                </b-navbar-nav>
+
+                <!-- Right aligned nav items -->
+                <b-navbar-nav class="ml-auto">
+                    <b-nav-item href="#" @click="handleLogin">{{ loginText }}</b-nav-item>
+                    <b-nav-item href="#" @click="signIn" v-if="loginStore.isLogin">회원가입</b-nav-item>
+                    
+                </b-navbar-nav>
+            </b-collapse>
+        </b-navbar>
+    </div>
+    <TeleportModal v-if="modalStore.modal.modalLogin">
+        <LoginModal @getUserInfo="getUserInfo" />
+    </TeleportModal>
+    <TeleportModal v-if="modalStore.modal.modalSignIn">
+        <SignInModal @postUserInfo="postUserInfo" />
+    </TeleportModal>
+    <router-view></router-view>
+</template>
+
+<script setup>
+import TeleportModal from './modal/TeleportModal.vue';
+import LoginModal from './modal/LoginModal.vue';
+import SignInModal from './modal/SignInModal.vue';
+
+import { ref } from 'vue';
+
+import { useModalStore } from '@/stores/modal';
+import { useListStore } from '@/stores/token';
+import { useLoginStore } from '@/stores/isLogin';
+
+import { postLogin, postSignIn } from '@/api/userApi.js';
+
+const modalStore = useModalStore();
+const listStore = useListStore();
+const loginStore = useLoginStore();
+
+function getUserInfo(email, password) {
+  const accessToken = postLogin(email, password)
+
+  listStore.addList({
+    accessToken: accessToken,
+  })
+
+  loginText.value = "로그아웃"
+  loginStore.changeStatus();
+}
+
+function postUserInfo(email, password, name, phoneNumber) {
+  const data = postSignIn(email, password, name, phoneNumber)
+
+  if (data.status == 400) {
+    alert("이메일 중복")
+  }
+  else {
+    alert("회원가입 성공")
+  }
+}
+
+function handleLogin() {
+  if (loginText.value == "로그인") {
+    modalStore.openModal('modalLogin');
+  }
+  else {
+    loginText.value = "로그인";
+    loginStore.changeStatus();
+  }
+}
+
+function signIn() {
+  modalStore.openModal('modalSignIn');
+}
+
+const loginText = ref("로그인");
+
+</script>
+
+<style>
+.ml-auto {
+    margin-left: auto!important;
+}
+</style>
