+++ frontend/src/pages/DataAnalysis.vue
@@ -1,0 +1,5 @@
+<template>
+    데이터 분석
+</template>
+<script setup>
+</script>
