+++ frontend/src/pages/Machine.vue
@@ -1,0 +1,9 @@
+<template>
+    <div>
+        머신러닝 페이지
+    </div>
+</template>
+
+<script setup>
+
+</script>
