+++ frontend/src/pages/Programming.vue
@@ -1,0 +1,5 @@
+<template>
+    프로그래밍
+</template>
+<script setup>
+</script>
